[#ICC-117] update after fn-app cleanup
