+++ src/services/__tests__/newMessagesService.test.ts
@@ -3,17 +3,21 @@
 import * as e from "express";
 import * as t from "io-ts";
 
+import * as TE from "fp-ts/TaskEither";
+
 import { mockedUser } from "../../__mocks__/user_mock";
 import NewMessageService from "../newMessagesService";
 import mockRes from "../../__mocks__/response";
 import { GetMessagesParameters } from "../../../generated/parameters/GetMessagesParameters";
-import { MessageStatusChange } from "../../../generated/io-api/MessageStatusChange";
-import { Change_typeEnum as Reading_Change_typeEnum } from "../../../generated/io-api/MessageStatusReadingChange";
-import { MessageStatusValueEnum } from "../../../generated/io-api/MessageStatusValue";
+import { MessageStatusChange } from "../../../generated/io-messages-api/MessageStatusChange";
+import { Change_typeEnum as Reading_Change_typeEnum } from "../../../generated/io-messages-api/MessageStatusReadingChange";
+import { MessageStatusValueEnum } from "../../../generated/io-messages-api/MessageStatusValue";
 import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";
-import { MessageStatusAttributes } from "../../../generated/io-api/MessageStatusAttributes";
-import { MessageStatusWithAttributes } from "../../../generated/io-api/MessageStatusWithAttributes";
+import { MessageStatusAttributes } from "../../../generated/io-messages-api/MessageStatusAttributes";
+import { MessageStatusWithAttributes } from "../../../generated/io-messages-api/MessageStatusWithAttributes";
 import { AppMessagesAPIClient } from "../../clients/app-messages.client";
+import { IPecServerClientFactoryInterface } from "../IPecServerClientFactory";
+import { IPecServerClient } from "../../clients/pecserver";
 
 const aValidMessageId = "01C3GDA0GB7GAFX6CCZ3FK3Z5Q" as NonEmptyString;
 const aPublicMessageParam = true;
@@ -103,6 +107,32 @@ const validApiMessageWithEnrichedDataResponse = {
   }
 };
 
+const validApiMessageResponseWithPrescriptionMetadata = {
+  status: 200,
+  value: {
+    message: {
+      content: {
+        markdown: aValidMarkdown,
+        prescription_data: {
+          iup: "12345678",
+          nre: "12345678",
+          prescriber_fiscal_code: "SPNDNL80R13C600R"
+        },
+        subject: aValidSubject
+      },
+      created_at: "2018-06-12T09:45:06.771Z",
+      fiscal_code: "LSSLCU79B24L219P",
+      id: "01CFSP4XYK3Y0VZTKHW9FKS1XM",
+      sender_service_id: "5a563817fcc896087002ea46c49a"
+    },
+    notification: {
+      email: "SENT",
+      webhook: "SENT"
+    },
+    status: "PROCESSED"
+  }
+};
+
 const emptyApiMessagesResponse = {
   status: 404
 };
@@ -139,6 +169,101 @@ const mockGetMessages = jest.fn();
 const mockGetMessage = jest.fn();
 const mockUpsertMessageStatus = jest.fn();
 
+// ------------------------
+// Legal message
+// ------------------------
+const aValidPecServerJwtToken = "aValidToken";
+
+const aValidLegalData = {
+  sender_mail_from: "test@legal.it",
+  has_attachment: false,
+  message_unique_id: "A_MSG_UNIQUE_ID"
+};
+
+const validApiMessageResponseWithLegalData = {
+  status: 200,
+  value: {
+    message: {
+      content: {
+        markdown: "a".repeat(81),
+        subject: aValidSubject,
+        legal_data: aValidLegalData
+      },
+      created_at: new Date(),
+      fiscal_code: "LSSLCU79B24L219P",
+      id: "01CFSP4XYK3Y0VZTKHW9FKS1XM",
+      sender_service_id: "5a563817fcc896087002ea46c49a"
+    },
+    notification: {
+      email: "SENT",
+      webhook: "SENT"
+    },
+    status: "PROCESSED"
+  }
+};
+
+const validApiLegalMessageResponse = {
+  status: 200,
+  value: {
+    cert_data: {
+      data: {
+        envelope_id: "anEnvelopeId",
+        msg_id: "<msgId@pec.it>",
+        receipt_type: "completa",
+        sender_provider: "aProvider",
+        timestamp: new Date()
+      },
+      header: {
+        object: "anObject",
+        recipients: "aRecipient@pec.it",
+        replies: "sender@pec.it",
+        sender: "sender@pec.it"
+      }
+    },
+    eml: {
+      attachments: [
+        {
+          content_type: "application/pdf",
+          id: "0",
+          name: "attachment_name",
+          url: "/messages/message_unique_id/attachments/0"
+        }
+      ],
+      html_content: "",
+      plain_text_content: "aPlainTextContent",
+      subject: "A Legal Subject"
+    }
+  }
+};
+
+const proxyLegalMessageResponse =
+  validApiMessageResponseWithLegalData.value.message;
+
+const mockGetLegalMessage = jest.fn();
+const mockGetLegalMessageAttachment = jest.fn();
+const mockPecServerApiClient: Partial<ReturnType<IPecServerClient>> = {
+  getMessage: mockGetLegalMessage,
+  getAttachmentBody: mockGetLegalMessageAttachment
+};
+const pecServerClientFactoryMock = {
+  getClient: jest.fn().mockImplementation(() => TE.of(mockPecServerApiClient))
+} as IPecServerClientFactoryInterface;
+
+const aValidAttachmentResponse = {
+  status: 200,
+  arrayBuffer: jest
+    .fn()
+    .mockImplementation(() => Promise.resolve(Buffer.from("anAttachment")))
+};
+
+const aBearerGenerator = jest
+  .fn()
+  .mockImplementation(() => TE.of(aValidPecServerJwtToken));
+
+// ----------------------------
+// Tests
+// ----------------------------
+
 beforeEach(() => {
   jest.clearAllMocks();
 });
@@ -155,7 +280,7 @@ describe("MessageService#getMessagesByUser", () => {
       return t.success(validApiMessagesResponse);
     });
 
-    const service = new NewMessageService(api);
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
 
     const res = await service.getMessagesByUser(mockedUser, mockParameters);
 
@@ -173,7 +298,7 @@ describe("MessageService#getMessagesByUser", () => {
       t.success(emptyApiMessagesResponse)
     );
 
-    const service = new NewMessageService(api);
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
 
     const res = await service.getMessagesByUser(mockedUser, mockParameters);
 
@@ -188,7 +313,7 @@ describe("MessageService#getMessagesByUser", () => {
       t.success(tooManyReqApiMessagesResponse)
     );
 
-    const service = new NewMessageService(api);
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
 
     const res = await service.getMessagesByUser(mockedUser, mockParameters);
 
@@ -198,7 +323,7 @@ describe("MessageService#getMessagesByUser", () => {
   it("returns an error if the getMessagesByUser API returns an error", async () => {
     mockGetMessages.mockImplementation(() => t.success(problemJson));
 
-    const service = new NewMessageService(api);
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
 
     const res = await service.getMessagesByUser(mockedUser, mockParameters);
     expect(mockGetMessages).toHaveBeenCalledWith({
@@ -212,7 +337,7 @@ describe("MessageService#getMessagesByUser", () => {
       t.success(invalidApiMessagesResponse)
     );
 
-    const service = new NewMessageService(api);
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
 
     const res = await service.getMessagesByUser(mockedUser, mockParameters);
     expect(mockGetMessages).toHaveBeenCalledWith({
@@ -226,7 +351,7 @@ describe("MessageService#getMessage", () => {
   it("returns a message from the API", async () => {
     mockGetMessage.mockImplementation(() => t.success(validApiMessageResponse));
 
-    const service = new NewMessageService(api);
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
 
     const res = await service.getMessage(
       mockedUser,
@@ -248,7 +373,7 @@ describe("MessageService#getMessage", () => {
       t.success(validApiMessageWithEnrichedDataResponse)
     );
 
-    const service = new NewMessageService(api);
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
 
     const res = await service.getMessage(mockedUser, getMessageParam);
 
@@ -263,10 +388,26 @@ describe("MessageService#getMessage", () => {
     });
   });
 
+  it("returns a message with attachments from the API", async () => {
+    mockGetMessage.mockImplementation(() =>
+      t.success(validApiMessageResponseWithPrescriptionMetadata)
+    );
+
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
+
+    const res = await service.getMessage(mockedUser, { id: aValidMessageId });
+
+    expect(mockGetMessage).toHaveBeenCalledWith({
+      fiscal_code: mockedUser.fiscal_code,
+      id: aValidMessageId
+    });
+    expect(res).toMatchSnapshot();
+  });
+
   it("returns an error if the getMessage API returns an error", async () => {
     mockGetMessage.mockImplementation(() => t.success(problemJson));
 
-    const service = new NewMessageService(api);
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
 
     const res = await service.getMessage(
       mockedUser,
@@ -284,7 +425,7 @@ describe("MessageService#getMessage", () => {
       t.success(invalidApiMessageResponse)
     );
 
-    const service = new NewMessageService(api);
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
 
     const res = await service.getMessage(
       mockedUser,
@@ -302,7 +443,7 @@ describe("MessageService#getMessage", () => {
       t.success(tooManyReqApiMessagesResponse)
     );
 
-    const service = new NewMessageService(api);
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
 
     const res = await service.getMessage(
       mockedUser,
@@ -310,6 +451,215 @@ describe("MessageService#getMessage", () => {
     );
 
     expect(res.kind).toEqual("IResponseErrorTooManyRequests");
+  });
+});
+
+describe("MessageService#getLegalMessage", () => {
+  it("returns a legal message from the API", async () => {
+    mockGetMessage.mockImplementation(async () =>
+      t.success(validApiMessageResponseWithLegalData)
+    );
+    mockGetLegalMessage.mockImplementation(async () =>
+      t.success(validApiLegalMessageResponse)
+    );
+
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
+
+    const res = await service.getLegalMessage(
+      mockedUser,
+      aValidMessageId,
+      aBearerGenerator
+    );
+
+    expect(mockGetMessage).toHaveBeenCalledWith({
+      fiscal_code: mockedUser.fiscal_code,
+      id: aValidMessageId
+    });
+    expect(res).toMatchObject({
+      kind: "IResponseSuccessJson",
+      value: proxyLegalMessageResponse
+    });
+  });
+  it("returns an error if the getMessage API returns an error", async () => {
+    mockGetMessage.mockImplementation(async () => t.success(problemJson));
+
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
+
+    const res = await service.getLegalMessage(
+      mockedUser,
+      aValidMessageId,
+      aBearerGenerator
+    );
+    expect(mockGetMessage).toHaveBeenCalledWith({
+      fiscal_code: mockedUser.fiscal_code,
+      id: aValidMessageId
+    });
+    expect(res.kind).toEqual("IResponseErrorInternal");
+  });
+
+  it("returns unknown response if the response from the getMessage API returns something wrong", async () => {
+    mockGetMessage.mockImplementation(async () =>
+      t.success(invalidApiMessageResponse)
+    );
+
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
+
+    const res = await service.getLegalMessage(
+      mockedUser,
+      aValidMessageId,
+      aBearerGenerator
+    );
+    expect(mockGetMessage).toHaveBeenCalledWith({
+      fiscal_code: mockedUser.fiscal_code,
+      id: aValidMessageId
+    });
+    expect(res.kind).toEqual("IResponseErrorInternal");
+  });
+
+  it("returns an error response if the response from the getMessage API returns a message without legal data", async () => {
+    mockGetMessage.mockImplementation(async () =>
+      t.success(validApiMessageResponse)
+    );
+
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
+
+    const res = await service.getLegalMessage(
+      mockedUser,
+      aValidMessageId,
+      aBearerGenerator
+    );
+    expect(mockGetMessage).toHaveBeenCalledWith({
+      fiscal_code: mockedUser.fiscal_code,
+      id: aValidMessageId
+    });
+    expect(res.kind).toEqual("IResponseErrorInternal");
+  });
+
+  it("returns an error response if bearer generator fails to generate a valid jwt", async () => {
+    mockGetMessage.mockImplementation(async () =>
+      t.success(validApiMessageResponse)
+    );
+
+    aBearerGenerator.mockImplementation(() =>
+      TE.left(new Error("Cannot generate jwt"))
+    );
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
+
+    const res = await service.getLegalMessage(
+      mockedUser,
+      aValidMessageId,
+      aBearerGenerator
+    );
+    expect(mockGetMessage).toHaveBeenCalledWith({
+      fiscal_code: mockedUser.fiscal_code,
+      id: aValidMessageId
+    });
+    expect(res.kind).toEqual("IResponseErrorInternal");
+  });
+
+  it("returns an error response if the response from the getLegalMessage API returns an error", async () => {
+    mockGetMessage.mockImplementation(async () =>
+      t.success(validApiLegalMessageResponse)
+    );
+
+    mockGetLegalMessage.mockImplementation(async () => t.success(problemJson));
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
+
+    const res = await service.getLegalMessage(
+      mockedUser,
+      aValidMessageId,
+      aBearerGenerator
+    );
+    expect(mockGetMessage).toHaveBeenCalledWith({
+      fiscal_code: mockedUser.fiscal_code,
+      id: aValidMessageId
+    });
+    expect(res.kind).toEqual("IResponseErrorInternal");
+  });
+});
+
+describe("MessageService#getLegalMessageAttachment", () => {
+  it("returns a legal message attachment from the API", async () => {
+    mockGetMessage.mockImplementation(async () =>
+      t.success(validApiMessageResponseWithLegalData)
+    );
+    mockGetLegalMessageAttachment.mockImplementationOnce(() =>
+      TE.of(aValidAttachmentResponse)
+    );
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
+
+    const res = await service.getLegalMessageAttachment(
+      mockedUser,
+      aValidMessageId,
+      aBearerGenerator,
+      aValidMessageId
+    );
+    expect(res).toMatchObject({
+      kind: "IResponseSuccessOctet",
+      value: aValidAttachmentResponse.arrayBuffer()
+    });
+  });
+
+  it("returns a response error if bearerGenerator fails to generate a valid jwt", async () => {
+    mockGetMessage.mockImplementation(async () =>
+      t.success(validApiMessageResponseWithLegalData)
+    );
+    mockGetLegalMessageAttachment.mockImplementationOnce(() =>
+      TE.of(aValidAttachmentResponse)
+    );
+
+    aBearerGenerator.mockImplementation(() =>
+      TE.left(new Error("Cannot generate jwt"))
+    );
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
+
+    const res = await service.getLegalMessageAttachment(
+      mockedUser,
+      aValidMessageId,
+      aBearerGenerator,
+      aValidMessageId
+    );
+    expect(res).toMatchObject({
+      kind: "IResponseSuccessOctet",
+      value: aValidAttachmentResponse.arrayBuffer()
+    });
+  });
+  it("returns an error if there are connectivity error on getLegalMessageAttachment API", async () => {
+    mockGetMessage.mockImplementation(async () =>
+      t.success(validApiMessageResponseWithLegalData)
+    );
+
+    mockGetLegalMessageAttachment.mockImplementationOnce(() =>
+      TE.left(new Error("Connection timeout"))
+    );
+
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
+
+    const res = await service.getLegalMessageAttachment(
+      mockedUser,
+      aValidMessageId,
+      aBearerGenerator,
+      aValidMessageId
+    );
+    expect(res.kind).toEqual("IResponseErrorInternal");
+  });
+  it("returns an error if the getLegalMessageAttachment API returns an error", async () => {
+    mockGetMessage.mockImplementation(async () =>
+      t.success(validApiMessageResponseWithLegalData)
+    );
+    mockGetLegalMessageAttachment.mockImplementationOnce(() =>
+      TE.left(new Error("Problem"))
+    );
+
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
+
+    const res = await service.getLegalMessageAttachment(
+      mockedUser,
+      aValidMessageId,
+      aBearerGenerator,
+      aValidMessageId
+    );
+    expect(res.kind).toEqual("IResponseErrorInternal");
   });
 });
 
@@ -339,7 +689,7 @@ describe("MessageService#upsertMessageStatus", () => {
       });
     });
 
-    const service = new NewMessageService(api);
+    const service = new NewMessageService(api, pecServerClientFactoryMock);
     const res = await service.upsertMessageStatus(
       mockedUser.fiscal_code,
       aValidMessageId as NonEmptyString,
@@ -387,7 +737,7 @@ describe("MessageService#upsertMessageStatus", () => {
         });
       });
 
-      const service = new NewMessageService(api);
+      const service = new NewMessageService(api, pecServerClientFactoryMock);
       const res = await service.upsertMessageStatus(
         mockedUser.fiscal_code,
         aValidMessageId as NonEmptyString,
